Fixed major bugs regarding "dependent: :destroy"
I found a serious bug where because of the placement of "dependent: :destroy" it would delete the account of the user. Even when I switched around the placement of it, it still gave me bugs so I think for now it would be a good idea not to have it and we'll deal with deleting all of the notes for a deleted account through other means.