Implemented public notes feature.
Now when editing or creating a new note there will be a checkbox where you can make the note public. If the note is public any one with the url can view the note. Only the author of the note will only ever have access to the "View Index", "Edit Note", and "Delete Note" buttons. Private notes will still redirect users that are not the author.